feat: use "HasCount" for counting items in a collection
