+++ Source/aweXpect/That/Collections/ThatEnumerable.AtMost.cs
@@ -22,15 +22,4 @@ public static partial class ThatEnumerable
 		this IThat<IEnumerable<string?>?> subject,
 		int maximum)
 		=> new(subject, EnumerableQuantifier.AtMost(maximum));
-
-	/// <summary>
-	///     Verifies that the collection has at most <paramref name="maximum" /> items.
-	/// </summary>
-	public static ItemsResult<AndOrResult<IEnumerable<TItem>, IThat<IEnumerable<TItem>?>>> AtMost<TItem>(
-		this IThatHas<IEnumerable<TItem>?> source,
-		int maximum)
-		=> new(new AndOrResult<IEnumerable<TItem>, IThat<IEnumerable<TItem>?>>(
-			source.ExpectationBuilder.AddConstraint((it, grammar)
-				=> new SyncCollectionCountConstraint<TItem>(it, EnumerableQuantifier.AtMost(maximum))),
-			source.ExpectSubject()));
 }
